tested Gamma and nbinom dists

+++ src/sportstradamus/train.py
@@ -1,4 +1,5 @@
 from sportstradamus.stats import StatsMLB, StatsNBA, StatsNHL, StatsNFL
+from sportstradamus.helpers import get_ev
 import pickle
 import importlib.resources as pkg_resources
 from sportstradamus import data
@@ -9,7 +10,7 @@ from sklearn.metrics import (
     accuracy_score,
     brier_score_loss,
     mean_tweedie_deviance,
-    mean_squared_log_error
+    log_loss
 )
 from scipy.stats import (
     norm,
@@ -34,9 +35,10 @@ from lightgbmlss.distributions.distribution_utils import DistributionClass
 @click.command()
 @click.option("--force/--no-force", default=False, help="Force update of all models")
 @click.option("--stats/--no-stats", default=False, help="Regenerate model reports")
+@click.option("--alt/--no-alt", default=False, help="Generate alternate model sets")
 @click.option("--league", type=click.Choice(["All", "NFL", "NBA", "MLB", "NHL"]), default="All",
               help="Select league to train on")
-def meditate(force, stats, league):
+def meditate(force, stats, league, alt):
 
     dist_params = {
         "stabilization": "None",
@@ -66,7 +68,6 @@ def meditate(force, stats, league):
 
     all_markets = {
         "NFL": [
-            "tds",
             "passing yards",
             "rushing yards",
             "receiving yards",
@@ -78,6 +79,7 @@ def meditate(force, stats, league):
             "passing tds",
             "rushing tds",
             "receiving tds",
+            "tds",
             "qb tds",
             "completions",
             "carries",
@@ -173,7 +175,7 @@ def meditate(force, stats, league):
             filename = "_".join([league, market]).replace(" ", "-") + ".mdl"
             filepath = pkg_resources.files(data) / filename
             if os.path.isfile(filepath) and not force:
-                if stats:
+                if stats or alt:
                     with open(filepath, 'rb') as infile:
                         filedict = pickle.load(infile)
                         model = filedict['model']
@@ -186,7 +188,7 @@ def meditate(force, stats, league):
             print(f"Training {league} - {market}")
             filename = "_".join([league, market]).replace(" ", "-")
             filepath = pkg_resources.files(data) / (filename + ".csv")
-            if os.path.isfile(filepath):
+            if os.path.isfile(filepath) and not force:
                 M = pd.read_csv(filepath, index_col=0)
                 M.to_csv(filepath)
             else:
@@ -255,10 +257,56 @@ def meditate(force, stats, league):
                             num_boost_round=n_rounds
                             )
 
+            elif alt:
+                y_train = np.ravel(y_train.to_numpy())
+
+                dtrain = lgb.Dataset(X_train, label=y_train)
+
+                if dist == "Gaussian":
+                    dist = "Gamma"
+                    # X_train = X_train[y_train > 1]
+                    # y_train = y_train[y_train > 1]
+                    y_train[y_train < 1] = 1
+                elif dist == "Poisson":
+                    dist = "NegativeBinomial"
+                else:
+                    continue
+
+                params = {
+                    "boosting_type": ["categorical", ["gbdt"]],
+                    "max_depth": ["int", {"low": 2, "high": 63, "log": False}],
+                    "num_leaves": ["int", {"low": 7, "high": 4095, "log": False}],
+                    "min_child_weight": ["float", {"low": 1e-2, "high": len(X)*.8/1000, "log": True}],
+                    "feature_fraction": ["float", {"low": 0.4, "high": 1.0, "log": False}],
+                    "bagging_fraction": ["float", {"low": 0.4, "high": 1.0, "log": False}],
+                    "bagging_freq": ["int", {"low": 1, "high": 1, "log": False}]
+                }
+
+                model = LightGBMLSS(distributions[dist])
+                opt_param = model.hyper_opt(params,
+                                            dtrain,
+                                            num_boost_round=999,
+                                            nfold=5,
+                                            early_stopping_rounds=20,
+                                            max_minutes=30,
+                                            n_trials=500,
+                                            silence=True,
+                                            )
+                opt_params = opt_param.copy()
+                n_rounds = opt_params["opt_rounds"]
+                del opt_params["opt_rounds"]
+
+                model.train(opt_params,
+                            dtrain,
+                            num_boost_round=n_rounds
+                            )
+
             X_test.loc[X_test["Line"] == 0, "Line"] = X_test.loc[X_test["Line"]
                                                                  == 0, "Avg10"].apply(np.ceil)-0.5
             X_test.loc[X_test["Line"] <= 0, "Line"] = 0.5
             prob_params = model.predict(X_test, pred_type="parameters")
+            prob_params.index = y_test.index
+            prob_params['result'] = y_test['Result']
             if dist == "Poisson":
                 under = poisson.cdf(
                     X_test["Line"], prob_params["rate"])
@@ -267,37 +315,40 @@ def meditate(force, stats, league):
                 y_proba = under - push/2
                 p = 1
                 ev = prob_params["rate"]
+                entropy = np.mean(
+                    np.abs(poisson.cdf(y_test["Result"], prob_params["rate"])-.5))
             elif dist == "Gaussian":
                 y_proba = norm.cdf(
                     X_test["Line"], prob_params["loc"], prob_params["scale"])
                 p = 0
                 ev = prob_params["loc"]
+                entropy = np.mean(
+                    np.abs(norm.cdf(y_test["Result"], prob_params["loc"], prob_params["scale"])-.5))
             elif dist == "Gamma":
                 y_proba = gamma.cdf(
                     X_test["Line"], prob_params["concentration"], scale=1/prob_params["rate"])
                 p = 2
                 ev = prob_params["concentration"]/prob_params["rate"]
+                entropy = np.mean(np.abs(gamma.cdf(
+                    y_test["Result"], prob_params["concentration"], scale=1/prob_params["rate"])-.5))
+                y_test.loc[y_test["Result"] == 0, 'Result'] = 1
             elif dist == "NegativeBinomial":
                 y_proba = nbinom.cdf(
                     X_test["Line"], prob_params["total_count"], prob_params["probs"])
                 p = 1
                 ev = prob_params["total_count"] * \
                     (1-prob_params["probs"])/prob_params["probs"]
+                entropy = np.mean(np.abs(norm.cdf(
+                    y_test["Result"], prob_params["total_count"], prob_params["probs"])-.5))
 
             y_proba = np.array([y_proba, 1-y_proba]).transpose()
             y_class = (y_test["Result"] >
                        X_test["Line"]).astype(int)
             y_class = np.ravel(y_class.to_numpy())
-            y_pred = (y_proba > .54).astype(int)
-            preco = precision_score(
-                (y_class == 1).astype(int), y_pred[:, 1])
-            precu = precision_score(
-                (y_class == 0).astype(int), y_pred[:, 0])
-            y_pred = y_pred[:, 1]
-            acc = accuracy_score(
-                y_class[np.max(y_proba, axis=1) > 0.54], y_pred[np.max(
-                    y_proba, axis=1) > 0.54]
-            )
+            y_pred = (y_proba > .5).astype(int)[:, 1]
+            mask = np.max(y_proba, axis=1) > 0.54
+            prec = precision_score(y_class[mask], y_pred[mask])
+            acc = accuracy_score(y_class[mask], y_pred[mask])
 
             bs = brier_score_loss(y_class, y_proba[:, 1], pos_label=1)
             bs0 = brier_score_loss(
@@ -308,27 +359,30 @@ def meditate(force, stats, league):
             dev0 = mean_tweedie_deviance(y_test, X_test["Line"], power=p)
             dev = 1 - dev/dev0
 
-            rmlse = mean_squared_log_error(y_test, ev, squared=False)
-            rmlse0 = mean_squared_log_error(
-                y_test, X_test["Line"], squared=False)
-            rmlse = 1 - rmlse/rmlse0
+            ll = log_loss(y_class, y_proba[:, 1])
+            ll0 = log_loss(y_class, 0.5*np.ones_like(y_class))
+            ll = 1 - ll/ll0
+
+            entropy0 = np.mean(np.abs(poisson.cdf(
+                y_test["Result"], X_test["Line"].apply(get_ev, args=(.5,)))-0.5))
+
+            entropy = 1 - entropy/entropy0
 
             filedict = {
                 "model": model,
                 "stats": {
                     "Accuracy": acc,
-                    "Precision_Over": preco,
-                    "Precision_Under": precu,
-                    "Brier Score Loss": bs,
+                    "Precision": prec,
+                    "Entropy": entropy,
+                    "Likelihood": ll,
+                    "Brier Score": bs,
                     "Deviance": dev,
-                    "RMLSE": rmlse
                 },
                 "params": params,
                 "distribution": dist
             }
 
             X_test['Result'] = y_test['Result']
-            X_test.reset_index(inplace=True, drop=True)
             if dist == "Poisson":
                 X_test['EV'] = prob_params['rate']
             elif dist == "Gaussian":
@@ -373,7 +427,7 @@ def report():
             market = name[1].replace("-", " ").replace(".mdl", "")
             dist = model["distribution"]
 
-            f.write(f" {league} {market} ".center(80, "="))
+            f.write(f" {league} {market} ".center(65, "="))
             f.write("\n")
             f.write(f" Distribution Model: {dist}\n")
             f.write(pd.DataFrame(model['stats'], index=[
